Deleted runbook US url test and improve DS runbook urls tes
Test for Upstream runbook urls is not relevant, this PR deletes it
and simplify the runbook downstream urls test.
